starting, moving to HPC

+++ cli/afni_task_subj.py
@@ -61,6 +61,7 @@ def submit_jobs(
     do_decon,
     decon_plan,
     do_blur,
+    do_clean,
 ):
     """Schedule work for single participant.
 
@@ -94,6 +95,8 @@ def submit_jobs(
         planned deconvolution with behavior: timing file mappings
     do_blur : bool
         [T/F] whether to blur as part of pre-processing
+    do_clean : bool
+        [T/F] whether to keep (F) or remove (T) intemediates
 
     Returns
     -------
@@ -149,6 +152,9 @@ def submit_jobs(
             print(f"Finished {subj}/{sess}/{task} with: \\n {{afni_data}}")
 
         # clean up
+        if not {do_clean}:
+            return
+
         shutil.rmtree(os.path.join("{afni_dir}", "{subj}", "{sess}", "sbatch_out"))
         clean_dir = os.path.join("{afni_dir}", "{subj}", "{sess}")
         clean_list = [
@@ -270,21 +276,24 @@ def get_args():
             """
         ),
     )
+    parser.add_argument(
+        "--clean",
+        action="store_false",
+        help=textwrap.dedent(
+            """\
+            Toggle of whether to remove intermediates.
+            Boolean (False if "--clean", else True,
+            True = intermediates will be removed).
+            """
+        ),
+    )
 
     required_args = parser.add_argument_group("Required Arguments")
     required_args.add_argument(
-        "-s",
-        "--session",
-        help="BIDS session str (ses-S2)",
-        type=str,
-        required=True,
+        "-s", "--session", help="BIDS session str (ses-S2)", type=str, required=True,
     )
     required_args.add_argument(
-        "-t",
-        "--task",
-        help="BIDS EPI task str (task-test)",
-        type=str,
-        required=True,
+        "-t", "--task", help="BIDS EPI task str (task-test)", type=str, required=True,
     )
     required_args.add_argument(
         "-c",
@@ -328,6 +337,7 @@ def main():
     task = args.task
     code_dir = args.code_dir
     do_blur = args.blur
+    do_clean = args.clean
 
     # set up
     log_dir = os.path.join(code_dir, "logs")
@@ -406,8 +416,7 @@ def main():
     # submit workflow.control_afni for each subject
     current_time = datetime.now()
     slurm_dir = os.path.join(
-        afni_dir,
-        f"""slurm_out/afni_{current_time.strftime("%y-%m-%d_%H:%M")}""",
+        afni_dir, f"""slurm_out/afni_{current_time.strftime("%y-%m-%d_%H:%M")}""",
     )
     if not os.path.exists(slurm_dir):
         os.makedirs(slurm_dir)
@@ -427,6 +436,7 @@ def main():
             value_dict["Decon"],
             value_dict["Decon_plan"],
             do_blur,
+            do_clean,
         )
         time.sleep(3)
         print(f"submit_jobs out: {h_out} \nsubmit_jobs err: {h_err}")
